feat: move users array
