+++ src/index.ts
@@ -1,20 +1,9 @@
-import http from 'http';
 import 'dotenv/config';
 
-import { findRoute } from './utils/utils';
+import { startProgram } from './startProgram';
+import { users } from './localDb/users';
+
 
 const PORT = process.env.PORT || 3000;
 
-const startProgram = () => {
-  const server = http.createServer((req, res) => {
-    const pathname = req.url || '';
-    const method = req.method || 'GET';
-    findRoute({ method, pathname, req, res });
-  });
-
-  server.listen(PORT, () => {
-    console.log(`Server running at http://localhost:${PORT}/`);
-  });
-};
-
-startProgram();
+startProgram(PORT, () => users);
